Revert Repetier hotend auto tune P, I, D constants.
I was thinking that while printing it is a different system than while resting, so perhaps the autotune P, I, D constants are not useful.  For the heated print bed of course, they are unaffected by the gradual printing process as the print slowly comes into contact. Even the heated print bed could have more heat moving into the part and affect the system.